rename remove to delete
